add storing of cook book

+++ src/main/java/todo/custom/cook/book/io/CookBookIO.java
@@ -17,13 +17,25 @@ public final class CookBookIO {
 	return Files.exists(cookBookFilePath);
     }
 
-    public CookBook get() throws IOException {
-	final String jsonContent = Files.readString(cookBookFilePath);
-	return jsonParser.fromJson(jsonContent, CookBook.class);
+    public CookBook get() {
+	try {
+	    final String jsonContent = Files.readString(cookBookFilePath);
+	    return jsonParser.fromJson(jsonContent, CookBook.class);
+	} catch (final IOException e) {
+	    throw new IllegalStateException("Unable to fetch existing cook book from file system", e);
+	}
     }
 
-    public void store(final CookBook cookBook) throws IOException {
-	final String jsonContent = jsonParser.toJson(cookBook);
-	Files.writeString(cookBookFilePath, jsonContent, StandardOpenOption.WRITE);
+    public void store(final CookBook cookBook) {
+	try {
+	    final String jsonContent = jsonParser.toJson(cookBook);
+	    if (cookBookExists()) {
+		Files.writeString(cookBookFilePath, jsonContent, StandardOpenOption.WRITE);
+	    } else {
+		Files.writeString(cookBookFilePath, jsonContent, StandardOpenOption.CREATE_NEW);
+	    }
+	} catch (final IOException e) {
+	    throw new IllegalStateException("Unabel to store cook book to file system", e);
+	}
     }
 }
